kasuari

Struct Expression

pub struct Expression {
    pub terms: Vec<Term>,
    pub constant: f64,
}

An expression that can be the left hand or right hand side of a constraint equation.

It is a linear combination of variables, i.e., a sum of variables weighted by coefficients, plus an optional constant.

expression = term_1 + term_2 + ... + term_n + constant

Fields
terms: Vec<Term>

The terms in the expression.

constant: f64

The constant in the expression.

Implementations

impl Expression

pub const fn new(terms: Vec<Term>, constant: f64) -> Expression

Create a new Expression.

expression = term_1 + term_2 + ... + term_n + constant

pub const fn from_constant(constant: f64) -> Expression

Constructs an expression that represents a constant without any terms

expression = constant

pub fn from_term(term: Term) -> Expression

Constructs an expression from a single term.

expression = term

pub const fn from_terms(terms: Vec<Term>) -> Expression

Constructs an expression from a terms

expression = term_1 + term_2 + ... + term_n

pub fn from_variable(variable: Variable) -> Expression

Constructs an expression from a variable

expression = variable