kasuari

Struct Strength

pub struct Strength(/* private fields */);

Implementations

impl Strength

pub const REQUIRED: Strength

The required strength for a constraint. This is the strongest possible strength.

pub const STRONG: Strength

A strong strength for a constraint. This is weaker than REQUIRED but stronger than MEDIUM.

pub const MEDIUM: Strength

A medium strength for a constraint. This is weaker than STRONG but stronger than WEAK.

pub const WEAK: Strength

A weak strength for a constraint. This is weaker than MEDIUM but stronger than 0.0.

pub const ZERO: Strength

The weakest possible strength for a constraint. This is weaker than WEAK.

pub const fn new(value: f64) -> Self

Create a new strength with the given value, clipped to the legal range (0.0, REQUIRED)

pub const fn create( strong: f64, medium: f64, weak: f64, multiplier: f64, ) -> Self

Create a constraint as a linear combination of STRONG, MEDIUM and WEAK strengths.

Each weight is multiplied by the multiplier, clamped to the legal range and then multiplied by the corresponding strength. The resulting strengths are then summed.

pub const fn value(&self) -> f64

The value of the strength

pub const fn add(self, rhs: Self) -> Self

Add two strengths together, clamping the result to the legal range

pub const fn sub(self, rhs: Self) -> Self

Subtract one strength from another, clipping the result to the legal range

pub const fn mul_f64(self, rhs: f64) -> Self

Multiply a strength by a scalar, clipping the result to the legal range

pub const fn mul_f32(self, rhs: f32) -> Self

Multiply a strength by a scalar, clipping the result to the legal range

pub const fn div_f64(self, rhs: f64) -> Self

Divide a strength by a scalar, clipping the result to the legal range

pub const fn div_f32(self, rhs: f32) -> Self

Divide a strength by a scalar, clipping the result to the legal range